+++ src/browserist/browser/click/button.py
@@ -8,7 +8,7 @@ from ...model.type.xpath import XPath
 from ..wait.for_element import wait_for_element
 
 
-def click_button(driver: object, xpath: str, timeout: int = timeout.DEFAULT) -> None:
+def click_button(driver: object, xpath: str, timeout: float = timeout.DEFAULT) -> None:
     xpath = XPath(xpath)
     wait_for_element(driver, xpath, timeout)
     try:
