+++ src/browserist/browser/mouse/hover.py
@@ -6,7 +6,7 @@ from ...model.type.xpath import XPath
 from ..wait.for_element import wait_for_element
 
 
-def mouse_hover(driver: object, xpath: str, timeout: int = timeout.DEFAULT) -> None:
+def mouse_hover(driver: object, xpath: str, timeout: float = timeout.DEFAULT) -> None:
     xpath = XPath(xpath)
     wait_for_element(driver, xpath, timeout)
     element = driver.find_element(By.XPATH, xpath)  # type: ignore
