+++ src/browserist/browser/tool/count_elements.py
@@ -5,7 +5,7 @@ from ...model.type.xpath import XPath
 from ..wait.for_element import wait_for_element
 
 
-def tool_count_elements(driver: object, xpath: str, timeout: int = timeout.DEFAULT) -> int:
+def tool_count_elements(driver: object, xpath: str, timeout: float = timeout.DEFAULT) -> int:
     xpath = XPath(xpath)
     wait_for_element(driver, xpath, timeout)
     elements: list[object] = driver.find_elements(By.XPATH, xpath)  # type: ignore
